Final polishing and bug fixes

+++ SpaceManager.java
@@ -14,17 +14,28 @@ public class SpaceManager implements Serializable {
     }
 
     public void deallocate(VirtualFile file){
+        if(file.isDeleted()){
+            throw new RuntimeException();
+        }
         allocator.deallocate(file);
     }
 
     public void deleteDirectory(Directory directory){
+        if(directory.isDeleted()){
+            throw new RuntimeException();
+        }
         Queue<Directory> directories = new LinkedList<Directory>();
         directories.add(directory);
         while(!directories.isEmpty()){
             directory = directories.poll();
             directory.deleteDirectory();
             for (VirtualFile file : directory.getFiles()) {
-                deallocate(file);
+                try{
+                    deallocate(file);
+                }
+                catch(Exception e){
+
+                }
             }
             for (Directory toBeDeleted : directory.getSubDirectories()) {
                 directories.add(toBeDeleted);
